Add componenent category emojis to refactor code
